Fix cl-loop destructuring
cl-loop doesn't technically support plist destructuring, but as long as
they are given in the same order, it should work.  And it does...except
when I run in the MELPA sandbox, it throws a (setting-constant :name)
error.  Why does it work in my personal config?  Who knows.  So I have
to change the keywords to _, which works fine, but...sigh.  Good thing I
tested in the sandbox.