WIP
Not sure if this is a good idea.  I want to make the most common case
the easiest.  Having the default order be at the very top seems okay,
until you want to move one section, to near the top, and then you also
have to set the order for a bunch of others.  So setting the default
order to something like 10 seems like it might be better, but it doesn't
seem to provide great results either.  I think I need to experiment more
to find the best default value.