Docs: Workaround for GitHub readme rendering
Setting org-comment-string buffer-locally is a nasty hack to work
around GitHub's org-ruby's HTML rendering, which does not respect
noexport tags.

Apparently, the only way to hide a tree from its output is to use the
COMMENT keyword, but that prevents Org from processing the export
options declared in it, which would require us to put all the options
at the top of the file, which is messy, and unlike org-manual.org,
which we are trying to imitate.

So since these file-local variables don't affect org-ruby, wet set
org-comment-string to an unused keyword, which prevents Org from
deleting this tree from the export buffer, which allows it to find the
export options in it.  And since org-export does respect the noexport
tag, the tree is excluded from the info page.