Tests: (test.org) Update for Org 9.2
Two issues:

1.  The org-get-repeat function now seems to expect habit entries to
have repeaters like ".+2d" and rejects ones like "+2d".  That causes
the org-habit-parse-todo function to raise an error.  So the habit
entry in this test data now has a period-prefixed repeater.

2.  Something changed in the way the Agenda formats inherited tags.
Previously, a tag that was both inherited and local was displayed as a
local tag, i.e. after the double-colon.  For example:

    TODO Blah blah  :inherited::inherited_and_local:local:

But now such a tag is displayed before the double-colon, as if it were
only inherited, e.g.:

    TODO Blah blah  :inherited:inherited_and_local::local:

This broke many stored test results in this project.  (I also didn't
even realize that this feature, to display inherited tags behind a
double-colon, existed.  It's handy, if you know what the double-colon
means.  I wonder if it's documented...)

Regarding the changed behavior in org-agenda.el: As best I can tell,
the function org-agenda-fix-displayed-tags is what implements this
functionality, and the only change to that function between
2017-02-10 (the Org version I'm still using in my personal Emacs
config) and now is c9dd7a6a9de34b8094ad6a419260fcc1214cc791, which
doesn't appear to have any effect on the way they are formatted.  So
I'm guessing the change is elsewhere, probably in the code that
applies "inherited" text properties to tag strings.

I don't know if this change is intentional or a bug.  It might be
worth looking into, although it is very minor.