Change/Fix: (--map-children) Use org-narrow-to-subtree
This fixes a weird bug in which certain tests run with the :children
selector would work interactively but fail when run from a script,
because the buffer would not be narrowed correctly--it would include
too much of the buffer.  I can't explain it at all, but this fixes
it.

I don't know if this bug manifested in normal use, so I'm not putting
it in the changelog.

Also, simplify the macro: it's not necessary to test the results
again, because the cond returns either nil or non-nil.