Use SCHEDULED/DEADLINE template for timestamps
The defgroup for SCHEDULED and DEADLINE are practically the same, save
for the special property they’re retrieving with ‘org-entry-get’.
However, the defgroup for matching timestamps is different, probably
due to an earlier conception.

For instance, compared to its SCHEDULED and DEADLINE equivalents, this
defgroup does not allow matching past or future timestamps with 'past
or 'future.

In order to address this issue, and for the sake of standardisation,
this commit reuses the template for SCHEDULED and DEADLINE, and only
modified the special property in ‘org-entry-get’.