Docs: Remove image from readme
I still fully support the statement intended by the Web page the flag
links to, which is intended to advocate non-political, non-partisan
participation in software projects, regardless of other participants'
views, ideologies, qualities, characteristics, etc.

Unfortunately, the flag's imagery is--quite sadly--commonly
misinterpreted to mean the opposite of what is intended.  The message
of liberty and equality is twisted and misused.

The flag's presence in some of my projects' documentation arose out of
a specific situation that happened in some Emacs-related projects
several years ago, when it seemed necessary to take a stand publicly
to support certain maintainers who were being treated unfairly.

Thankfully, that situation has long since passed.  So it doesn't seem
necessary to keep this public statement prominently visible anymore.