Tests: Update for newer Emacs/ERT versions
These tests used to all work and pass, but since some changes to Emacs
27 and/or 28, they failed in various ways, some of which seem to be
dependent on whether they are byte-compiled (which, with makem.sh,
they are by default).

These changes seem to fix the test suite with regard to the changes in
Emacs and ERT.  However, a few issues remain:

1.  Somehow, the cl-letf* form that temporarily redefines
format-time-string is not properly restoring the binding after the
test body, and ERT itself gives an error when it tries to call the
temporary binding, format-time-string-orig, after running all of the
tests.  I can't explain this, but at least these changes get us closer
to a fully functional test suite again.

2.  Some of the changes in recent Org versions have included minor
differences in Agenda output, which some tests' results need to be
updated for.  This can be dealt with later.