defcustom: org-super-agenda-ignored-agenda-names
List of agenda names on which to avoid using Org Super Agenda.  The
agenda name is as specified by the local variable `org-agenda-name' of
the agenda to be ignored by Org Super Agenda.

The function `org-super-agenda--filter-finalize-entries' was modified
in order to not finalize the entries of the agenda when the agenda
being processed match one of the strings defined in the custom
variable `org-super-agenda-ignored-agenda-names' defined by this
commit.

Example:

> (custom-set-variables
>   '(org-super-agenda-ignored-agenda-names
>     '("^TAGS" "^SEARCH$" "^TODO$")))